+++ src/components/Plugins/index.ts
@@ -3,6 +3,8 @@ export * from './Balena';
 export * from './Base64Image';
 export * from './Calendar';
 export * from './DataManipulation';
+export * from './DynamicText';
 export * from './Environment';
 export * from './Plugins';
 export * from './RSSAtom';
+export * from './Static';