+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 1.4.0 (IN PROGRESS)
+## 1.4.0 (2022-10-29)
 
 ### Features / Enhancements
 
@@ -8,7 +8,8 @@
 - Add Compatibility Check Workflow (#25)
 - Update to Grafana 9.1.6 (#26)
 - Update to Grafana 9.2.2 (#29)
-- Update GitHub Actions to use Node 16 (#30)
+- Update CI to Node 16 and Synchronize with Release workflow (#30)
+- Add Dynamic Text, Static Data Source (#31)
 
 ## 1.3.0 (2022-09-10)